# Conflicts:
#	bin/fr/ecp/IS1220/group5/project/Coordinate.class
#	bin/fr/ecp/IS1220/group5/project/Courier.class
#	bin/fr/ecp/IS1220/group5/project/Customer.class
#	bin/fr/ecp/IS1220/group5/project/DefaultDiscount.class
#	bin/fr/ecp/IS1220/group5/project/Discount.class
#	bin/fr/ecp/IS1220/group5/project/Financial.class
#	bin/fr/ecp/IS1220/group5/project/Item.class
#	bin/fr/ecp/IS1220/group5/project/ItemType.class
#	bin/fr/ecp/IS1220/group5/project/Manager.class
#	bin/fr/ecp/IS1220/group5/project/Meal.class
#	bin/fr/ecp/IS1220/group5/project/MealType.class
#	bin/fr/ecp/IS1220/group5/project/MyFoodoraSystem.class
#	bin/fr/ecp/IS1220/group5/project/Order.class
#	bin/fr/ecp/IS1220/group5/project/Restaurant.class
#	bin/fr/ecp/IS1220/group5/project/User.class
#	bin/fr/ecp/IS1220/group5/project/UserNotFoundException.class
#	src/Test/CoordinateTest.java
#	src/fr/ecp/IS1220/group5/project/MyFoodoraSystem.java
#	src/fr/ecp/IS1220/group5/project/Order.java

+++ src/fr/ecp/IS1220/group5/project/Manager.java
@@ -5,10 +5,16 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Observable;
 import java.util.Observer;
-
+/**
+ * <b>The class that represents a Manager</b>
+ * It extends the <b>User</b> abstract class.
+ *
+ * @see User
+ */
 public class Manager extends User implements Observer{
 	
 	private String surname;
@@ -72,21 +78,21 @@ public class Manager extends User implements Observer{
 		}
 	}
 
-	public void changeServiceFee(double servicefee){
+	public void changeServiceFee(BigDecimal servicefee){
 		Financial financial = null;
 		financial = retrieveFinancial();
 		financial.service_fee = servicefee;
 		saveFinancial(financial);
 	}
 
-	public void changeMarkup_percentage(double markup_percentage){
+	public void changeMarkup_percentage(BigDecimal markup_percentage){
 		Financial financial = null;
 		financial = retrieveFinancial();
 		financial.markup_percentage = markup_percentage;
 		saveFinancial(financial);
 	}
 
-	public void changeDelivery_cost(double delivery_cost){
+	public void changeDelivery_cost(BigDecimal delivery_cost){
 		Financial financial = null;
 		financial = retrieveFinancial();
 		financial.delivery_cost = delivery_cost;
